##  Overview

This PR implements a comprehensive reorganization of the test suite and unification of GitHub Actions workflows to improve consistency, performance, and maintainability.

##  Test Architecture Reorganization

### New Layered Test Structure:
- **Unit Tests** (	ests/unit/): Fast tests with no external dependencies
  - Run on multiple Python versions (3.10, 3.11, 3.12)
  - ~32 tests covering core functionality
- **Local Tests** (	ests/local/): Tests requiring local server startup
  - Run on Python 3.11 (matching Copilot environment)
  - Server process management and SSE endpoint testing
- **Integration Tests** (	ests/integration/): Tests requiring external services
  - Run on Python 3.11, only on push to master
  - PowerBI/XMLA connectivity validation

### Performance Improvements:
-  Optimized 	est_server_stays_running from ~24s to ~5s using process monitoring
-  All 32 tests pass with improved execution speed
-  Layered execution prevents unnecessary test runs

##  GitHub Actions Workflow Unification

### New Composite Action:
- **Location**: .github/actions/setup-powerbi-environment/
- **Purpose**: Unified environment setup across all contexts (Docker, Copilot, CI)
- **Key Features**:
  - Uses shared install_dotnet_adomd.sh script for consistency
  - Standardized Python 3.11 across environments
  - Configurable installation modes (--system vs --user)

### Workflow Improvements:
- **Copilot Setup Steps**: Simplified to use composite action
- **Tests Workflow**: Layered execution with dependency flow
- **Environment Consistency**: Same setup between Copilot and CI/CD
- **Merge Commit Handling**: Removed exclusions - tests now run after PR merges

##  Key Benefits

1. ** Environment Consistency**: Copilot and CI use identical environments
2. ** Performance**: Optimized test execution with smart layering  
3. ** Maintainability**: Eliminated code duplication via composite actions
4. ** Reliability**: Better test coverage with targeted execution strategy

##  Files Changed

- .github/actions/setup-powerbi-environment/action.yml  New composite action
- .github/workflows/copilot-setup-steps.yml  Simplified using composite action
- .github/workflows/tests.yml  Layered test execution strategy
- Previous test reorganization: 	ests/unit/, 	ests/local/, 	ests/integration/

##  Testing

- All 32 tests pass in new layered structure
- Performance improvements validated (24s  5s for slow test)
- Workflow syntax validated (lint warnings are normal for missing secrets)

##  Breaking Changes

None - this is purely an internal reorganization that improves the development experience without affecting the external API.

---

**Ready for review and testing!**